build udev lib bins statically
This patch extracts both stratis-base32-decode and stratis-str-cmp from
stratis-util. They are now two distinct binaries.
Also, when built, they are built statically.

```
$ ldd target/x86_64-unknown-linux-gnu/release/stratis-base32-decode
        statically linked
```
Related: https://github.com/stratis-storage/stratisd/issues/3195
